Removed all cyclic dependencies.
