+++ lib/asn1/api.js
@@ -1,4 +1,5 @@
-var asn1 = require('../asn1');
+var decoders = require('./decoders');
+var encoders = require('./encoders');
 var inherits = require('inherits');
 
 var api = exports;
@@ -39,7 +40,7 @@ Entity.prototype._createNamed = function createNamed(base) {
 Entity.prototype._getDecoder = function _getDecoder(enc) {
   // Lazily create decoder
   if (!this.decoders.hasOwnProperty(enc))
-    this.decoders[enc] = this._createNamed(asn1.decoders[enc]);
+    this.decoders[enc] = this._createNamed(decoders[enc]);
   return this.decoders[enc];
 };
 
@@ -50,7 +51,7 @@ Entity.prototype.decode = function decode(data, enc, options) {
 Entity.prototype._getEncoder = function _getEncoder(enc) {
   // Lazily create encoder
   if (!this.encoders.hasOwnProperty(enc))
-    this.encoders[enc] = this._createNamed(asn1.encoders[enc]);
+    this.encoders[enc] = this._createNamed(encoders[enc]);
   return this.encoders[enc];
 };
