+++ lib/asn1/base/buffer.js
@@ -1,5 +1,5 @@
 var inherits = require('inherits');
-var Reporter = require('../base').Reporter;
+var Reporter = require('./reporter').Reporter;
 var Buffer = require('buffer').Buffer;
 
 function DecoderBuffer(base, options) {
