+++ lib/asn1/base/node.js
@@ -1,5 +1,5 @@
-var Reporter = require('../base').Reporter;
-var EncoderBuffer = require('../base').EncoderBuffer;
+var Reporter = require('./reporter').Reporter;
+var EncoderBuffer = require('./buffer').EncoderBuffer;
 var assert = require('minimalistic-assert');
 
 // Supported tags
