+++ lib/asn1/constants/der.js
@@ -1,4 +1,4 @@
-var constants = require('../constants');
+var utils = require('./utils');
 
 exports.tagClass = {
   0: 'universal',
@@ -6,7 +6,7 @@ exports.tagClass = {
   2: 'context',
   3: 'private'
 };
-exports.tagClassByName = constants._reverse(exports.tagClass);
+exports.tagClassByName = utils._reverse(exports.tagClass);
 
 exports.tag = {
   0x00: 'end',
@@ -39,4 +39,4 @@ exports.tag = {
   0x1d: 'charstr',
   0x1e: 'bmpstr'
 };
-exports.tagByName = constants._reverse(exports.tag);
+exports.tagByName = utils._reverse(exports.tag);
