+++ lib/asn1/constants/index.js
@@ -1,19 +1,5 @@
 var constants = exports;
+var utils = require('./utils');
 
-// Helper
-constants._reverse = function reverse(map) {
-  var res = {};
-
-  Object.keys(map).forEach(function(key) {
-    // Convert key to integer if it is stringified
-    if ((key | 0) == key)
-      key = key | 0;
-
-    var value = map[key];
-    res[value] = key;
-  });
-
-  return res;
-};
-
+constants._reverse = utils._reverse;
 constants.der = require('./der');
