+++ lib/asn1/decoders/der.js
@@ -1,11 +1,8 @@
 var inherits = require('inherits');
 
-var asn1 = require('../../asn1');
-var base = asn1.base;
-var bignum = asn1.bignum;
-
-// Import DER constants
-var der = asn1.constants.der;
+var base = require('../base');
+var bignum = require('bn.js');
+var der = require('../constants/der');
 
 function DERDecoder(entity) {
   this.enc = 'der';
