+++ lib/asn1/decoders/pem.js
@@ -1,7 +1,5 @@
 var inherits = require('inherits');
 var Buffer = require('buffer').Buffer;
-
-var asn1 = require('../../asn1');
 var DERDecoder = require('./der');
 
 function PEMDecoder(entity) {
