+++ lib/asn1/encoders/der.js
@@ -1,12 +1,10 @@
 var inherits = require('inherits');
 var Buffer = require('buffer').Buffer;
-
-var asn1 = require('../../asn1');
-var base = asn1.base;
-var bignum = asn1.bignum;
+var base = require('../base');
+var constants = require('../constants');
 
 // Import DER constants
-var der = asn1.constants.der;
+var der = constants.der;
 
 function DEREncoder(entity) {
   this.enc = 'der';
